Calendar: Move generic utility functions to separate file
This keeps generator.ts limited to only contain actual logic (since this file is long & complex enough as it is)
